Move messages and actions to statusMessages file
